+++ src/Service/Cron/CronQuery.php
@@ -10,6 +10,6 @@ trait CronQuery
     private function getDeleteCronJobQuery() : string
     {
         return "DELETE FROM cron_job
-WHERE " . $this->ilias_database->like("component", ilDBConstants::T_TEXT, "flux_ilias_rest_%%") . " AND " . $this->ilias_database->like("component", ilDBConstants::T_TEXT, "%%_helper_plugin");
+WHERE " . $this->ilias_database->like("component", ilDBConstants::T_TEXT, "flux\\_ilias\\_rest\\_%") . " AND " . $this->ilias_database->like("component", ilDBConstants::T_TEXT, "%\\_helper\\_plugin");
     }
 }