+++ src/Service/ObjectConfig/ObjectConfigQuery.php
@@ -10,7 +10,7 @@ trait ObjectConfigQuery
     private function getDeleteObjectConfigsQuery() : string
     {
         return "DELETE FROM container_settings
-WHERE " . $this->ilias_database->like("keyword", ilDBConstants::T_TEXT, $this->getObjectConfigContainerSettingsPrefix() . "%%");
+WHERE " . $this->ilias_database->like("keyword", ilDBConstants::T_TEXT, str_replace("_", "\\_", $this->getObjectConfigContainerSettingsPrefix()) . "%");
     }